feat: move auth from utils to inpi folder
